requirements.txt
discord.py
Pillow
requests
unidecode
flask
